SUPERNOVA: Divide GameManager into 2 classes
GameManager got divided into a base class, that will be used
for both parts and a derived class specific to only the first
part.
